[ksqlDb.RestApi.Client]: extracted KSqlTypeTranslator from CreateEntity.  CreateEntityStatement renamed to EntityInfo. ksqlDB types are no longer automatically capitalized.
