+++ docs/breaking_changes.md
@@ -4,6 +4,7 @@
 - `TypeProperties` - removed generic type argument and added a base type `EntityProperties`. Additionally, changed from a class to a record type.
 - the above change affected the argument for the `IKSqlDbRestApiClient.CreateTypeAsync<T>(TypeProperties typeProperties)`
 - `IEntityCreationProperties` was replaced with `IEntityProperties`
+- ksqlDB types are no longer **automatically capitalized**
 
 Updated package references:
 ```xml
