+++ ksqlDb.RestApi.Client/KSql/RestApi/Generators/TypeGenerator.cs
@@ -7,12 +7,12 @@ using static ksqlDB.RestApi.Client.KSql.RestApi.Enums.IdentifierEscaping;
 
 namespace ksqlDB.RestApi.Client.KSql.RestApi.Generators;
 
-internal class TypeGenerator : CreateEntityStatement
+internal sealed class TypeGenerator : EntityInfo
 {
   internal string Print<T>(TypeProperties properties)
   {
     StringBuilder stringBuilder = new();
-    var typeName = EntityProvider.GetFormattedName<T>(properties, EscapeName).ToUpper();
+    var typeName = EntityProvider.GetFormattedName<T>(properties, EscapeName);
     stringBuilder.Append($"CREATE TYPE {typeName} AS STRUCT<");
 
     PrintProperties<T>(stringBuilder, properties.IdentifierEscaping);
@@ -30,9 +30,9 @@ internal class TypeGenerator : CreateEntityStatement
     {
       var type = GetMemberType(memberInfo);
 
-      var ksqlType = CreateEntity.KSqlTypeTranslator(type, escaping);
+      var ksqlType = KSqlTypeTranslator.Translate(type, escaping);
 
-      var columnDefinition = $"{EscapeName(memberInfo.Name, escaping)} {ksqlType}{CreateEntity.ExploreAttributes(memberInfo, type)}";
+      var columnDefinition = $"{EscapeName(memberInfo.Name, escaping)} {ksqlType}{KSqlTypeTranslator.ExploreAttributes(memberInfo, type)}";
       ksqlProperties.Add(columnDefinition);
     }
