+++ ksqlDb.RestApi.Client/KSql/RestApi/Statements/CreateEntity.cs
@@ -3,96 +3,13 @@ using System.Text;
 using ksqlDB.RestApi.Client.Infrastructure.Extensions;
 using ksqlDB.RestApi.Client.KSql.Query.Context;
 using ksqlDB.RestApi.Client.KSql.RestApi.Enums;
-using ksqlDB.RestApi.Client.KSql.RestApi.Extensions;
 using ksqlDb.RestApi.Client.KSql.RestApi.Parsers;
-using ksqlDB.RestApi.Client.KSql.RestApi.Statements.Annotations;
 using static System.String;
-using DateTime = System.DateTime;
 
 namespace ksqlDB.RestApi.Client.KSql.RestApi.Statements;
 
-internal sealed class CreateEntity : CreateEntityStatement
+internal sealed class CreateEntity : EntityInfo
 {
-  internal static string KSqlTypeTranslator(Type type, IdentifierEscaping escaping = IdentifierEscaping.Never)
-  {
-    var ksqlType = Empty;
-
-    if (type == typeof(byte[]))
-      ksqlType = KSqlTypes.Bytes;
-    else if (type.IsArray)
-    {
-      var elementType = KSqlTypeTranslator(type.GetElementType(), escaping);
-
-      ksqlType = $"ARRAY<{elementType}>";
-    }
-    else if (type.IsDictionary())
-    {
-      Type[] typeParameters = type.GetGenericArguments();
-
-      var keyType = KSqlTypeTranslator(typeParameters[0], escaping);
-      var valueType = KSqlTypeTranslator(typeParameters[1], escaping);
-
-      ksqlType = $"MAP<{keyType}, {valueType}>";
-    }
-    else if (type == typeof(string))
-      ksqlType = KSqlTypes.Varchar;
-    else if (type == typeof(Guid))
-      ksqlType = KSqlTypes.Varchar;
-    else if (type.IsOneOfFollowing(typeof(int), typeof(int?), typeof(short), typeof(short?)))
-      ksqlType = KSqlTypes.Int;
-    else if (type.IsOneOfFollowing(typeof(long), typeof(long?)))
-      ksqlType = KSqlTypes.BigInt;
-    else if (type.IsOneOfFollowing(typeof(double), typeof(double?)))
-      ksqlType = KSqlTypes.Double;
-    else if (type.IsOneOfFollowing(typeof(bool), typeof(bool?)))
-      ksqlType = KSqlTypes.Boolean;
-    else if (type == typeof(decimal))
-      ksqlType = KSqlTypes.Decimal;
-    else if (type == typeof(DateTime))
-      ksqlType = KSqlTypes.Date;
-    else if (type == typeof(TimeSpan))
-      ksqlType = KSqlTypes.Time;
-    else if (type == typeof(DateTimeOffset))
-      ksqlType = KSqlTypes.Timestamp;
-    else if (!type.IsGenericType && type.TryGetAttribute<StructAttribute>() != null)
-    {
-      var ksqlProperties = GetProperties(type, escaping);
-
-      ksqlType = $"STRUCT<{Join(", ", ksqlProperties)}>";
-    }
-    else if (!type.IsGenericType && (type.IsClass || type.IsStruct()))
-    {
-      ksqlType = type.Name.ToUpper();
-    }
-    else if (type.IsEnum)
-      ksqlType = KSqlTypes.Varchar;
-    else
-    {
-      Type elementType = null;
-
-      if (type.IsGenericType && type.GetGenericTypeDefinition() == typeof(IEnumerable<>))
-        elementType = type.GetGenericArguments()[0];
-      else
-      {
-        var enumerableInterfaces = type.GetEnumerableTypeDefinition().ToArray();
-
-        if (enumerableInterfaces.Length > 0)
-        {
-          elementType = enumerableInterfaces[0].GetGenericArguments()[0];
-        }
-      }
-
-      if (elementType != null)
-      {
-        string ksqlElementType = KSqlTypeTranslator(elementType, escaping);
-
-        ksqlType = $"ARRAY<{ksqlElementType}>";
-      }
-    }
-
-    return ksqlType;
-  }
-
   private readonly StringBuilder stringBuilder = new();
 
   internal string Print<T>(StatementContext statementContext, EntityCreationMetadata metadata, bool? ifNotExists)
@@ -119,34 +36,6 @@ internal sealed class CreateEntity : CreateEntityStatement
     return stringBuilder.ToString();
   }
 
-  internal static string ExploreAttributes(MemberInfo memberInfo, Type type)
-  {
-    if (type == typeof(decimal))
-    {
-      var decimalMember = memberInfo.TryGetAttribute<DecimalAttribute>();
-
-      if (decimalMember != null)
-        return $"({decimalMember.Precision},{decimalMember.Scale})";
-    }
-
-    if (type.IsArray)
-    {
-      var headersAttribute = memberInfo.TryGetAttribute<HeadersAttribute>();
-
-      if (headersAttribute != null)
-      {
-        const string header = " HEADER";
-
-        if (IsNullOrEmpty(headersAttribute.Key))
-          return $"{header}S";
-
-        return $"{header}('{headersAttribute.Key}')";
-      }
-    }
-
-    return Empty;
-  }
-
   private void PrintProperties<T>(StatementContext statementContext, EntityCreationMetadata metadata)
   {
     var ksqlProperties = new List<string>();
@@ -155,10 +44,10 @@ internal sealed class CreateEntity : CreateEntityStatement
     {
       var type = GetMemberType(memberInfo);
 
-      var ksqlType = KSqlTypeTranslator(type, metadata.IdentifierEscaping);
+      var ksqlType = KSqlTypeTranslator.Translate(type, metadata.IdentifierEscaping);
 
       var columnName = IdentifierUtil.Format(memberInfo, metadata.IdentifierEscaping);
-      string columnDefinition = $"\t{columnName} {ksqlType}{ExploreAttributes(memberInfo, type)}";
+      string columnDefinition = $"\t{columnName} {ksqlType}{KSqlTypeTranslator.ExploreAttributes(memberInfo, type)}";
 
       columnDefinition += TryAttachKey(statementContext.KSqlEntityType, memberInfo);
 
@@ -166,24 +55,6 @@ internal sealed class CreateEntity : CreateEntityStatement
     }
 
     stringBuilder.AppendLine(Join($",{Environment.NewLine}", ksqlProperties));
-  }
-
-  internal static IEnumerable<string> GetProperties(Type type, IdentifierEscaping escaping)
-  {
-    var ksqlProperties = new List<string>();
-
-    foreach (var memberInfo in Members(type, false))
-    {
-      var memberType = GetMemberType(memberInfo);
-
-      var ksqlType = KSqlTypeTranslator(memberType, escaping);
-
-      string columnDefinition = $"{memberInfo.Format(escaping)} {ksqlType}{ExploreAttributes(memberInfo, memberType)}";
-
-      ksqlProperties.Add(columnDefinition);
-    }
-
-    return ksqlProperties;
   }
 
   private void PrintCreateOrReplace<T>(StatementContext statementContext, EntityCreationMetadata metadata)
